+++ data/duke.txt
@@ -1,3 +1,3 @@
-E | 0 | slap | Aug 23 2022 9:00AM
 T | 0 | sleep
 E | 0 | poop | Aug 22 2022 4:00PM
+T | 0 | sleep
