+++ src/main/java/duke/DukeException.java
@@ -1,6 +1,8 @@
+package duke;
+
 /**
- * The DukeException Class represents
- * exceptions unique to Duke.
+ * The duke.DukeException Class represents
+ * exceptions unique to duke.Duke.
  */
 public class DukeException extends Exception {
     public DukeException(String description) {
