+++ src/main/java/duke/Parser.java
@@ -1,7 +1,8 @@
+package duke;
 
+import duke.commands.*;
 
 public class Parser {
-
 
     public static Command parse(String userInput) throws DukeException {
         String[] strings = userInput.split(" ");
