+++ src/main/java/duke/Storage.java
@@ -1,3 +1,10 @@
+package duke;
+
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.ToDo;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
@@ -6,8 +13,8 @@ import java.util.ArrayList;
 import java.util.Scanner;
 
 /**
- * The Storage Class represents the
- * storage for Duke.
+ * The duke.Storage Class represents the
+ * storage for duke.Duke.
  */
 public class Storage {
