+++ src/main/java/duke/Ui.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.task.Task;
+
 import java.util.ArrayList;
 import java.util.Scanner;
 public class Ui {
@@ -16,7 +20,7 @@ public class Ui {
                 + "| |_| | |_| |   <  __/\n"
                 + "|____/ \\__,_|_|\\_\\___|\n";
         System.out.println("Hello from\n" + logo);
-        System.out.println("Hello! I'm Duke \nWhat can I do for you?");
+        System.out.println("Hello! I'm duke.Duke \nWhat can I do for you?");
     }
 
     public void showLine() {
