+++ src/main/java/duke/commands/ByeCommand.java
@@ -1,3 +1,9 @@
+package duke.commands;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class ByeCommand extends Command {
 
     public static final String COMMAND_WORD = "bye";
