+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public abstract class Command {
