+++ src/main/java/duke/commands/DeadlineCommand.java
@@ -1,7 +1,14 @@
-public class EventCommand extends Command {
-    public static final String COMMAND_WORD = "event";
+package duke.commands;
+
+import duke.*;
+import duke.task.Deadline;
+import duke.task.Task;
+
+public class DeadlineCommand extends Command {
+
+    public static final String COMMAND_WORD = "deadline";
     private String[] userDescription;
-    public EventCommand(String[] userDescription) {
+    public DeadlineCommand(String[] userDescription) {
         this.userDescription = userDescription;
     }
 
@@ -13,13 +20,12 @@ public class EventCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         try {
-            Task event = new Event(userDescription[0], userDescription[1]);
-            tasks.addTask(event);
+            Task deadline = new Deadline(userDescription[0], userDescription[1]);
+            tasks.addTask(deadline);
             storage.save(tasks.getTaskList());
-            ui.showAddTask(event, tasks);
+            ui.showAddTask(deadline, tasks);
         } catch (ArrayIndexOutOfBoundsException e) {
             throw new DukeException("Please use /at to specify event time.");
         }
-
     }
 }
